+++ hspec-core/src/Test/Hspec/Core/Formatters.hs
@@ -1,4 +1,6 @@
+-- |
+-- Deprecated\: Use "Test.Hspec.Core.Formatters.V1" instead.
 module Test.Hspec.Core.Formatters
-{-# DEPRECATED "Use \"Test.Hspec.Core.Formatters.V1\" instead." #-}
+-- {-# DEPRECATED "Use \"Test.Hspec.Core.Formatters.V1\" instead." #-}
 (module V1) where
 import           Test.Hspec.Core.Formatters.V1 as V1