+++ CHANGES.markdown
@@ -1,5 +1,6 @@
 ## Changes in next
   - Add `--fail-on`.  This subsums `--fail-on-focused` and `--fail-on-pending`.
+  - Add `--fail-on=empty` (#650)
 
 ## Changes in 2.10.4
   - Make sure that whitespace-only diff output is colorized properly (#660)