+++ CHANGES.markdown
@@ -1,6 +1,9 @@
 ## Changes in next
   - Add `--fail-on`.  This subsums `--fail-on-focused` and `--fail-on-pending`.
   - Add `--fail-on=empty` (#650)
+  - Revert deprecation of `Test.Hspec.Formatters` and
+    `Test.Hspec.Core.Formatters`.  Instead the module documentation now says
+    that they are deprecated in favor of `Test.Hspec.Core.Formatters.V1`.
 
 ## Changes in 2.10.4
   - Make sure that whitespace-only diff output is colorized properly (#660)